feat: display edge obj count in group
