+++ frontend/src/components/KnowledgeConfigs.jsx
@@ -1,18 +1,25 @@
 import { Box, Menu, MenuButton, IconButton, Typography, Divider, Stack, Button, Select, Option, Dropdown } from "@mui/joy";
 import ArchitectureOutlinedIcon from '@mui/icons-material/ArchitectureOutlined';
 import useConfigStore from "../store/useConfigStore";
+import useStatusStore from "../store/useStatusStore";
+import useSelectorStore from "../store/useSelectorStore";
 
-export default function InteractionConfigurations() {
-    const nodeSizeMetric = useConfigStore(state => state.nodeSizeMetric);
-    const setNodeSizeMetric = useConfigStore(state => state.setNodeSizeMetric);
-    const edgeNotationMetric = useConfigStore(state => state.edgeNotationMetric);
-    const setEdgeNotationMetric = useConfigStore(state => state.setEdgeNotationMetric);
-    const clearInteractionGraphConfig = useConfigStore(state => state.clearInteractionGraphConfig);
 
-    const nodeSizeMetrics = ['none', 'objectCount', 'objectTypeCount'];
-    const nodeSizeMetricLabels = ['None', 'Total Object Count', 'Object Type Diversity'];
-    const edgeNotationMetrics = ['none', 'objectType', 'totalObjectCount', 'averageFlowTime'];
-    const edgeNotationMetricLabels = ['None', 'Object Type', 'Total Object Count', 'Average Flow Time'];
+export default function KnowledgeConfigs() {
+    const showingNodeType = useConfigStore(state => state.showingNodeType);
+    const setShowingNodeType = useConfigStore(state => state.setShowingNodeType);
+    const nodeSharing = useConfigStore(state => state.nodeSharing);
+    const setNodeSharing = useConfigStore(state => state.setNodeSharing);
+    const clearKnowledgeGraphConfig = useConfigStore(state => state.clearKnowledgeGraphConfig);
+
+    const mode = useStatusStore(state => state.mode);
+    const focusingNode = useSelectorStore(state => state.focusingNode);
+    const disabled = mode === 'knowledge' && focusingNode;
+
+    const showingNodeTypes = ['both', 'objectType', 'activity'];
+    const showingNodeTypeLabels = ['Both', 'Object Type', 'Event Type'];
+    const nodeSharingStatus = ['both', 'shared', 'nonshared'];
+    const nodeSharingStatusLabels = ['Both', 'Shared', 'Non-shared'];
 
     const Panel = () => (
         <Box sx={{ m: 2, width: 328 }}>
@@ -20,53 +27,53 @@ export default function InteractionConfigurations() {
                 Graph Configurations
             </Typography>
             <Typography level="body-sm" mb={2}>
-                Configure for the interaction graph.
+                Configure the knowledge graph.
             </Typography>
             <Divider sx={{ mb: 2 }} />
             <Stack direction="row" alignItems="center" justifyContent="space-between" mb={1}>
                 <Typography level="title-md">
-                    Node Size
+                    Showing Node Types
                 </Typography>
                 <Select 
-                    sx={{ width: '12rem' }} 
+                    sx={{ width: '10rem' }} 
                     slotProps={{ listbox: { disablePortal: true }}}
-                    defaultValue='none'
-                    value={nodeSizeMetric}
-                    onChange={(e, newValue) => setNodeSizeMetric(newValue)}
+                    defaultValue='both'
+                    value={showingNodeType}
+                    onChange={(e, newValue) => setShowingNodeType(newValue)}
                 >
-                    {nodeSizeMetrics.map((item, index) => (
+                    {showingNodeTypes.map((item, index) => (
                         <Option key={item} value={item}>
-                            {nodeSizeMetricLabels[index]}
+                            {showingNodeTypeLabels[index]}
                         </Option>
                     ))}
                 </Select>
             </Stack>
             <Typography level="body-sm" mb={2}>
-                Select a metric to determine the node sizes.
+                Configure which node types are displayed on the graph.
             </Typography>
             <Stack direction="row" alignItems="center" justifyContent="space-between" mb={1}>
                 <Typography level="title-md">
-                    Edge Annotation
+                    Node Share Status
                 </Typography>
                 <Select 
-                    sx={{ width: '12rem' }} 
+                    sx={{ width: '10rem' }} 
                     slotProps={{ listbox: { disablePortal: true }}}
-                    defaultValue='none'
-                    value={edgeNotationMetric}
-                    onChange={(e, newValue) => setEdgeNotationMetric(newValue)}
+                    defaultValue='both'
+                    value={nodeSharing}
+                    onChange={(e, newValue) => setNodeSharing(newValue)}
                 >
-                    {edgeNotationMetrics.map((item, index) => (
+                    {nodeSharingStatus.map((item, index) => (
                         <Option key={item} value={item}>
-                            {edgeNotationMetricLabels[index]}
+                            {nodeSharingStatusLabels[index]}
                         </Option>
                     ))}
                 </Select>
             </Stack>
             <Typography level="body-sm" mb={5}>
-                Select a metric to assign edge annotations.
+                Configure whether to display only shared nodes, only non-shared nodes, or all nodes on the graph. 
             </Typography>
             <Stack direction="row" alignItems="center">
-                <Button level="solid" onClick={() => clearInteractionGraphConfig()}>
+                <Button level="solid" onClick={() => clearKnowledgeGraphConfig()}>
                     Reset
                 </Button>
             </Stack>
@@ -76,6 +83,7 @@ export default function InteractionConfigurations() {
     return (
         <Dropdown>
             <MenuButton
+                disabled={disabled}
                 slots={{ root: IconButton }}
                 slotProps={{ root: { variant: 'outlined', color: 'neutral' } }}
             >
